requirements.txt
# nanobot requirements.txt

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Logging
loguru==0.7.2

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# JSON
orjson==3.9.10

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.12.1
ruff==0.1.9
mypy==1.8.0

# Development
ipython==8.19.0
